Fix shell creation in XP edge-case scenario
Today I found a scenario where spawning shells on Windows XP SP0 wasn't
working. This wasn't happening on Windows XP SP2, so it could be specific
to SP0.

The user was an administrative user, but running `shell` from the
Meterpreter prompt was resulting in the following error:

    meterpreter > shell
    [-] stdapi_sys_process_execute: Operation failed: 1314

This commit supports a FORCE flag which falls back to a non-impersonated
shell if the user specifies it.