Extra debug info, force adsi limit
The ADSI API is supposed to support the notion of only returning
a fixed number of results from AD when we specify the
ADS_SEARCHPREF_SIZE_LIMIT value in the search parameters, but it
doesn't appear to do that (at least not my Win2k setup).

This change includes a bit more debugging output and keeps track of
the number of rows processed so that the maximum site is actually
honoured.