Updated to Mimikatz commit e85d87f6bb2e0afa5c57f22c73d12b8e25e84c9f
This comes with some other changes as well which include code changes that
were probably made to the kerberos code prior to moving to github. As a result,
the actual changset was lots because the history of mimikatz wasn't imported
into github, and the change history is no longer available on google code.